refactor load data separate from save
