+++ modules/save_game.rb
@@ -16,32 +16,4 @@ module SaveGame
       file.puts(data.to_json)
     end
   end
-
-  def self.load_games_from_json
-    games = []
-    File.open('data/games.json', 'r') do |file|
-      file.each_line do |line|
-        data = JSON.parse(line)
-
-        author_data = data['author']
-        author = Author.new(author_data['first_name'], author_data['last_name'])
-
-        genre_data = data['genre']
-        genre = Genre.new(genre_data['name'])
-
-        label_data = data['label']
-        label = Label.new(title: label_data['name'], color: label_data['color'])
-        game = Game.new(data['publish_date'],
-                        data['multiplayer'],
-                        data['last_played_at'],
-                        archived: data['archived'],
-                        author: author,
-                        genre: genre,
-                        label: label,
-                        id: data['id'])
-        games << game
-      end
-    end
-    games
-  end
 end
